+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - feat: add `box3.union` to compute the union of two Box3s
 - feat: add `box3.size` to compute the size or dimensions of a Box3
 - feat: add `ray3.create`, `ray3.set`, `ray3.copy`, `ray3.fromSegment`
+- feat: add `box3.expandByMargin` to expand a Box3 by a uniform margin in all directions
 
 ## 0.0.6